Insert missing barlines
Unfortunately, this is harder to do in our new engine then the old
engine because we don't have a way of adding a measure while validating,
so we need to do this outside the validation process.  This is by design
as it allows us to reason about models, validation, and layout more easily.

This will be made efficient in the future:

 - we should save the status of measures not affected by inserting new
   measures (first, we need to finish up the validation memoization)
 - we should add all measures at once (right now, we're restarting from
   scratch for every measure added -- there's no reason this is
   necessary)

Test Plan: lilypond-regression 01b